Basic config loading with nconf
Using nconf to load up a file of config. For a single file ths is
undoubtedly overkill but once we get into multiple files and hierarchies
and overwriting it's going to save a lot of headaches.

Added two rather nasty tests (do not recommend or condone writing tests
like this) that are basically just present to ensure that Travis
complains if someone accidentally deletes/renames the master config or
commits with a master config that doesn't parse.

Also correctly wired the test script in package.json, closing #1.